+++ src/test/java/roomescape/dao/ReservationDaoTest.java
@@ -1,22 +1,20 @@
-package roomescape;
+package roomescape.dao;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.dao.ReservationDao;
-import roomescape.dao.ReservationTimeDao;
-import roomescape.dao.ThemeDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @JdbcTest
 class ReservationDaoTest {
@@ -36,22 +34,22 @@ class ReservationDaoTest {
         reservationTimeDao = new ReservationTimeDao(jdbcTemplate);
         themeDao = new ThemeDao(jdbcTemplate);
         ReservationTime reservationTime = reservationTimeDao.save(
-            new ReservationTime(
-                LocalTime.of(10, 5)
-            )
+                new ReservationTime(
+                        LocalTime.of(10, 5)
+                )
         );
         Theme savedTheme = themeDao.save(
-            new Theme(
-                "탈출 테스트",
-                "탈출하는 내용",
-                "abc.jpg"
-            )
+                new Theme(
+                        "탈출 테스트",
+                        "탈출하는 내용",
+                        "abc.jpg"
+                )
         );
         Reservation reservation = new Reservation(
-            "두리",
-            LocalDate.of(2025, 10, 5),
-            reservationTime,
-            savedTheme
+                "두리",
+                LocalDate.of(2025, 10, 5),
+                reservationTime,
+                savedTheme
         );
         savedReservation = reservationDao.save(reservation);
     }
@@ -60,23 +58,23 @@ class ReservationDaoTest {
     @DisplayName("Reservation 객체를 저장한다")
     void saveReservation() {
         Reservation reservation = new Reservation(
-            "두리",
-            LocalDate.of(2025, 10, 5),
-            new ReservationTime(
-                LocalTime.of(10, 5)
-            ),
-            new Theme(
-                "탈출",
-                "탈출하는 내용",
-                "abc.jpg"
-            )
+                "두리",
+                LocalDate.of(2025, 10, 5),
+                new ReservationTime(
+                        LocalTime.of(10, 5)
+                ),
+                new Theme(
+                        "탈출",
+                        "탈출하는 내용",
+                        "abc.jpg"
+                )
         );
         Reservation savedReservation = reservationDao.save(reservation);
 
         assertAll(
-            () -> assertThat(savedReservation.getName()).isEqualTo("두리"),
-            () -> assertThat(savedReservation.getDate()).isEqualTo(LocalDate.of(2025, 10, 5)),
-            () -> assertThat(savedReservation.getTime().getStartAt()).isEqualTo(LocalTime.of(10, 5))
+                () -> assertThat(savedReservation.getName()).isEqualTo("두리"),
+                () -> assertThat(savedReservation.getDate()).isEqualTo(LocalDate.of(2025, 10, 5)),
+                () -> assertThat(savedReservation.getTime().getStartAt()).isEqualTo(LocalTime.of(10, 5))
         );
     }
